move policy code to subdir
