+++ app/containers/LayerInfo/policy/CountryPositionSymbol.js
