+++ app/containers/LayerInfo/policy/charts.js
